Fix bug duplicate entry in product_supplier
When you duplicate a product from the backoffice in the "All shops" context, there is an SQL error because the result of the query contains the entries for each shop. To fix this bug, simply make a "GROUP BY id_product_attribute" at the end of the request to avoid having a duplicate in the result array.